+++ gaclen/examples/quad/shaders.rs
@@ -1,11 +1,11 @@
 pub mod vertex {
-	vulkano_shaders::shader!{
+	gaclen_shader::shader!{
 		ty: "vertex",
 		path: "examples/quad/shader.vert",
 	}
 }
 pub mod fragment {
-	vulkano_shaders::shader!{
+	gaclen_shader::shader!{
 		ty: "fragment",
 		path: "examples/quad/shader.frag",
 	}