+++ gaclen/src/graphics/pass.rs
@@ -1,0 +1,7 @@
+//! Infrastructure containing device configuration used for computations.
+
+mod graphical_pass;
+mod builder;
+
+pub use graphical_pass::*;
+pub use builder::GraphicalPassBuilder;